requirements.txt
telethon==1.34.0
feedparser==6.0.10
websocket-client==1.7.0
python-telegram-bot==20.7
asyncio-mqtt==0.16.1
colorama==0.4.6
rich==13.7.0
python-dotenv==1.0.0
requests==2.31.0
PySocks==1.7.1
